requirements.txt
discord.py
selenium
webdriver-manager
pyperclip
